requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
autopep8